Finish record client button.
The record client button records the data passed in by the user. There is one problem that needs to be looked at, which is the gender RadioGroup. I do not know how to work with it because we use binding when referring to the resources. Everything else seems to be recorded successfully.