Add camera button to new client page.
Added a camera button to the new client page and allows the user to take a photo; displays the image that is taken to the image view below the camera button.

Need to revisit this code and look at other tutorials to see if there is any better way to structure the code. Probably need to refactor a bit (function and variable names). Also, there is something wrong regarding the camera permissions.