Refactor new client code.
Started moving API calls to decouple the backend logic from the view. Referenced JohnnyD's branch for the implementation. So far only moved the API call for creating a client disability but will shortly refactor the rest of the code. Incrementally testing the code to see if it crashes so I can identify what is the source of the crash.